TASK-3_1: Refactor codes structure
