+++ lib/infrastructure/blockchain/nft_repository.dart
@@ -4,11 +4,11 @@ import 'dart:typed_data';
 import 'package:http/http.dart';
 
 import 'package:hnotes/domain/blockchain/dtos/nft_metadata_dto.dart';
-import 'package:hnotes/infrastructure/local_storage/files/folder_repository.dart';
+import 'package:hnotes/infrastructure/local_storage/files/nft_file_repository.dart';
 import 'package:hnotes/infrastructure/blockchain/base_blockchain_repository.dart';
 
 class NftRepository extends BaseBlockchainRepository {
-  FolderRepository _folderRepository = new FolderRepository();
+  NftFileRepository _nftFileRepository = new NftFileRepository();
 
   Future<NftMetaDataDto> getNFTMetadata(
       String contractAddress, int tokenId, String tokenType) async {
@@ -25,7 +25,7 @@ class NftRepository extends BaseBlockchainRepository {
     // Reference: https://stackoverflow.com/questions/52299112/flutter-download-an-image-from-url
     Response response = await get(Uri.parse(url));
     Uint8List content = response.bodyBytes;
-    String savedFileName = await _folderRepository.saveBytesFile(content, "$fileName.png");
+    String savedFileName = await _nftFileRepository.saveBytesFile(content, "$fileName.png");
     return savedFileName;
   }
 }
