+++ lib/infrastructure/local_storage/files/folder_repository.dart
@@ -1,32 +1,14 @@
 import 'dart:io';
-import 'dart:typed_data';
 import 'package:path_provider/path_provider.dart';
 
 class FolderRepository {
-  final String imagesFolderName = "images";
-  final String importedDataFolderName = "data";
-
-  Future<bool> createFoldersWhenLunch() async {
-    await createFolderInAppDocDir(imagesFolderName);
-    await createFolderInAppDocDir(importedDataFolderName);
-    return true;
-  }
-
-  Future<bool> deleteNft(String fileName) async {
-    await deleteFile(importedDataFolderName, fileName);
-    await deleteFile(imagesFolderName, fileName);
-    return true;
-  }
-
   Future<String> createFolderInAppDocDir(String folderNameParam) async {
     // App Document Directory + folder name
     final Directory appDocDirFolder = await folderUnderAppDir(folderNameParam);
 
-    // if folder already exists return path
     if (await appDocDirFolder.exists()) {
       return appDocDirFolder.path;
     } else {
-      // if folder not exists create folder and then return its path
       final Directory appDocDirNewFolder = await appDocDirFolder.create(recursive: true);
       return appDocDirNewFolder.path;
     }
@@ -47,18 +29,6 @@ class FolderRepository {
     return allFiles;
   }
 
-  Future<String> saveStringFile(String contents, String fileName) async {
-    File newFile = await _createFile(fileName, importedDataFolderName);
-    newFile.writeAsStringSync(contents);
-    return newFile.path;
-  }
-
-  Future<String> saveBytesFile(Uint8List bytes, String fileName) async {
-    File newFile = await _createFile(fileName, imagesFolderName);
-    newFile.writeAsBytesSync(bytes);
-    return newFile.path;
-  }
-
   Future<bool> deleteFile(String folderName, String fileName) async {
     String filePath = await _formFileName(fileName, folderName);
     File file = File(filePath);
@@ -67,10 +37,8 @@ class FolderRepository {
   }
 
   Future<Directory> folderUnderAppDir(String folderName) async {
-    // Get this App Document Directory
     final Directory appDocDir = await getApplicationDocumentsDirectory();
 
-    // App Document Directory + folder name
     final Directory appDocDirFolder = Directory('${appDocDir.path}/$folderName/');
 
     return appDocDirFolder;
@@ -83,7 +51,7 @@ class FolderRepository {
     return filePathAndName;
   }
 
-  Future<File> _createFile(String fileName, String folderNameParam) async {
+  Future<File> createFile(String fileName, String folderNameParam) async {
     String newFilePathAndName = await _formFileName(fileName, folderNameParam);
     File newFile = new File(newFilePathAndName);
     return newFile;
